Ensure RemoteAnimationTargets.unfilteredApps is not null
Public field RemoteAnimationTargets.unfilteredApps is accessed in
various locations in the code, also outside of the owning class. It is
generally assumed that it is a valid, non-null array.

In RemoteAnimationTargets constructor, if parameter apps is null,
unfilteredApps was not getting initialized at all. Initialize it with an
empty array instead.

Issue: FP3-A13#137
Change-Id: I3c185988819f8dc60035451bbbc2b9108ab59855